강의 관련 API 검증 및 에러 수정

+++ src/main/java/jiwoo/newKlasProject/DTO/Lecture/LectureScheduleDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class LectureScheduleDTO {
     private String lnum;
-    private int week;
+    private Long week;
     private String contents;
     private String method;
 }
